CORDA-3471: Address comments by @mnesbit - Relocate `CordaTransactionSupport` to `core` - Create a lighter version of transaction - `VaultTransaction` that gives access to `session` object only.
