+++ node/src/main/kotlin/net/corda/node/services/rpc/CheckpointDumper.kt
@@ -42,8 +42,7 @@ import net.corda.node.internal.NodeStartup
 import net.corda.node.services.api.CheckpointStorage
 import net.corda.node.services.statemachine.*
 import net.corda.node.utilities.JVMAgentUtil.getJvmAgentProperties
-import net.corda.nodeapi.internal.persistence.CordaPersistence
-import net.corda.nodeapi.internal.persistence.CordaTransactionSupport
+import net.corda.core.node.services.vault.CordaTransactionSupport
 import net.corda.serialization.internal.CheckpointSerializeAsTokenContextImpl
 import net.corda.serialization.internal.withTokenContext
 import java.nio.file.Path
