+++ node/src/main/kotlin/net/corda/notary/experimental/raft/RaftTransactionCommitLog.kt
@@ -30,7 +30,7 @@ import net.corda.core.utilities.contextLogger
 import net.corda.core.utilities.debug
 import net.corda.node.services.transactions.PersistentUniquenessProvider
 import net.corda.node.utilities.AppendOnlyPersistentMap
-import net.corda.nodeapi.internal.persistence.CordaTransactionSupport
+import net.corda.core.node.services.vault.CordaTransactionSupport
 import net.corda.nodeapi.internal.persistence.currentDBSession
 import net.corda.serialization.internal.CordaSerializationEncoding
 import java.time.Clock
